Add fix

+++ src/main/java/ru/yandex/practicum/filmorate/dao/LikeDbStorage.java
@@ -4,7 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
+import ru.yandex.practicum.filmorate.exceptions.DuplicateException;
+import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.storage.LikeStorage;
+
+import java.util.List;
 
 @Slf4j
 @Repository
@@ -14,7 +18,8 @@ public class LikeDbStorage implements LikeStorage {
     private final JdbcTemplate jdbc;
 
     @Override
-    public void addLike(Integer filmId, Integer userId) {
+//    public void addLike(Integer filmId, Integer userId) {
+    public void   addLike(Integer userId, Integer filmId) {
         if (!isFilmExists(filmId)) {
             throw new IllegalArgumentException("Фильма с id " + filmId + " не существует.");
         }
@@ -22,6 +27,11 @@ public class LikeDbStorage implements LikeStorage {
         if (isLikeAlreadyAdded(filmId, userId)) {
             return;
         }
+        if (isLikeAlreadyAdded(filmId, userId)) {
+            log.warn("Attempt to add existing like for film ID {} and user ID {}", filmId, userId);
+            throw new DuplicateException("Like from user ID " + userId + " for film ID " + filmId + " is existing.");
+        }
+
         String sql = "INSERT INTO film_likes (user_id, film_id) VALUES (?, ?)";
         jdbc.update(sql, userId, filmId);
     }
@@ -39,12 +49,23 @@ public class LikeDbStorage implements LikeStorage {
     }
 
     @Override
-    public void removeLike(Integer filmId, Integer userId) {
-        if (!isLikeAlreadyAdded(filmId, userId)) {
-            return;
+    public void removeLike(Integer userId, Integer filmId) {
+        if (!isFilmExists(filmId)) {
+            throw new IllegalArgumentException("Фильма с id " + filmId + " не существует.");
         }
-        log.info("Удаляем лайк для фильма с id " + filmId + " от пользователя с id " + userId);
+        if (!isLikeAlreadyAdded(filmId, userId)) {
+            log.warn("Attempt to remove non-existing like for film ID {} and user ID {}", filmId, userId);
+            throw new NotFoundException("Like from user ID " + userId + " for film ID " + filmId + " not found.");
+        }
+
+        log.info("Removing like for film ID {} from user ID {}", filmId, userId);
         String sql = "DELETE FROM film_likes WHERE user_id = ? AND film_id = ?";
         jdbc.update(sql, userId, filmId);
+    }
+
+    @Override
+    public List<Integer> getLikedFilmIds(Integer userId) {
+        String sql = "SELECT film_id FROM film_likes WHERE user_id = ?";
+        return jdbc.query(sql, (rs, rowNum) -> rs.getInt("film_id"), userId);
     }
 }
